[libtooling][biniou] fix integers too large for ocaml
Test Plan:
The compilation of objc models in infer fails without this:

  cd ~/infer
  make -C infer clang

Reviewers: mathieubaudet

Reviewed By: mathieubaudet

Subscribers: martinoluca

Differential Revision: https://phabricator.fb.com/D2360597

Tasks: 7792959

Signature: t1:2360597:1440005090:0f4162ca0d9bf5673878767e409ba12bc46c41c4

+++ libtooling/atdlib/ATDWriter.h
@@ -609,7 +609,7 @@ namespace ATDWriter {
       write32(x);
     }
 
-    void writeUvint(size_t x) {
+    void writeUvint(unsigned int x) {
       while (x > 127) {
         write8(x | 128);
         x >>= 7;
